printouts

+++ NXTservices/utils/NXTutils.h
@@ -773,7 +773,7 @@ char *submit_AM(CURL *curl_handle,char *recipient,struct NXT_AMhdr *ap,char *ref
    // printf("in submit_AM\n");
     memset(hexbytes,0,sizeof(hexbytes));
     init_hexbytes(hexbytes,(void *)ap,len);
-    sprintf(cmd,"%s=sendMessage&secretPhrase=%s&recipient=%s&message=%s&deadline=%u%s&feeNQT=%lld",_NXTSERVER,NXTACCTSECRET,recipient,hexbytes,deadline,reftxid!=0?reftxid:"",MIN_NQTFEE);
+    sprintf(cmd,"%s=sendMessage&secretPhrase=%s&recipient=%s&message=%s&deadline=%u%s&feeNQT=%lld",_NXTSERVER,NXTACCTSECRET,recipient,hexbytes,deadline,reftxid!=0?reftxid:"",(long long)MIN_NQTFEE);
     //printf("submit_AM.(%s)\n",cmd);
     jsonstr = issue_NXTPOST(curl_handle,cmd);
     printf("back from issue_NXTPOST.(%s) %p\n",jsonstr,curl_handle);
